requirements.txt
absl-py==1.1.0
accelerate==0.30.1
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
clip==1.0
cycler==0.11.0
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
einops==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fonttools==4.35.0
frozenlist==1.3.0
fsspec==2024.5.0
ftfy==6.1.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
huggingface-hub==0.23.2
idna==3.3
importlib-metadata==4.11.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.4
Markdown==3.3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.16
networkx==3.2.1
numpy==1.21.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
open-clip-torch==2.24.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
peft==0.11.1
pexpect==4.9.0
Pillow==9.1.1
prompt-toolkit==3.0.43
protobuf==3.19.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.8
safetensors==0.4.3
scipy==1.7.3
sentencepiece==0.2.0
simsimd==1.4.0
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.9.10
tokenizers==0.15.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.9.1
torchvision==0.14.1+cu116
tqdm==4.66.4
traitlets==5.14.1
transformers==4.36.2
triton==2.3.0
typing_extensions==4.12.0
uform==3.0.2
urllib3==1.26.9
wcwidth==0.2.13
Werkzeug==2.1.2
xxhash==3.4.1
yarl==1.7.2
zipp==3.8.0
